requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF
pyshp
pyproj
google-genai
python-dotenv
